HMAN-66 Refactor to validator package
