+++ src/main/java/uk/gov/hmcts/reform/hmc/Application.java
@@ -8,7 +8,7 @@ import org.springframework.scheduling.annotation.EnableAsync;
 @SpringBootApplication
 @EnableFeignClients
 @EnableAsync
-@SuppressWarnings("HideUtilityClassConstructor") // Spring needs a constructor, it's not a utility class
+@SuppressWarnings("HideUtilityClassConstructor") // Spring needs a constructor, it is not a utility class
 public class Application {
 
     public static void main(final String[] args) {
