+++ src/main/java/uk/gov/hmcts/reform/hmc/service/PartiesNotifiedServiceImpl.java
@@ -6,6 +6,7 @@ import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.reform.hmc.repository.HearingRepository;
 import uk.gov.hmcts.reform.hmc.repository.HearingResponseRepository;
+import uk.gov.hmcts.reform.hmc.validator.HearingIdValidator;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -13,7 +14,7 @@ import java.util.List;
 @Service
 @Component
 @Slf4j
-public class PartiesNotifiedServiceImpl extends HearingValidation implements PartiesNotifiedService {
+public class PartiesNotifiedServiceImpl extends HearingIdValidator implements PartiesNotifiedService {
 
     private final HearingResponseRepository hearingResponseRepository;
